move DelayStats out, change fields to atomic and make the instance static on RtpSender
this will aggregate the stats across all RtpSenders.  this also means
that it will no longer be printed as part of the pipeline stats, but the
idea is that it will be queried from elsewhere and reported somewhere
(e.g. the debug api)
