+++ src/main/kotlin/org/jitsi/nlj/stats/PacketDelayStats.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright @ 2018 - present 8x8, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jitsi.nlj.stats
+
+import org.jitsi.nlj.PacketInfo
+import java.util.concurrent.atomic.AtomicLong
+
+class DelayStats {
+    private val totalDelayMs = AtomicLong(0)
+    private val totalPackets = AtomicLong(0)
+    val averageDelay: Double
+        get() = totalDelayMs.get() / totalPackets.get().toDouble()
+    val maxDelayMs = AtomicLong(0)
+
+    fun addPacket(packetInfo: PacketInfo) {
+        val delayMs = if (packetInfo.receivedTime > 0) {
+            System.currentTimeMillis() - packetInfo.receivedTime
+        } else {
+            -1
+        }
+        if (delayMs >= 0) {
+            totalDelayMs.addAndGet(delayMs)
+            if (delayMs > maxDelayMs.get()) {
+                maxDelayMs.set(delayMs)
+            }
+            totalPackets.incrementAndGet()
+        }
+    }
+}
